Refactor album sidebar into modular components
