+++ src/iPhoto/gui/ui/delegates/__init__.py
@@ -1,0 +1,5 @@
+"""Delegate utilities for GUI widgets."""
+
+from .album_sidebar_delegate import AlbumSidebarDelegate, BranchIndicatorController
+
+__all__ = ["AlbumSidebarDelegate", "BranchIndicatorController"]
