+++ src/iPhoto/gui/ui/menus/__init__.py
@@ -1,0 +1,5 @@
+"""Context menu helpers for GUI widgets."""
+
+from .album_sidebar_menu import AlbumSidebarContextMenu, show_context_menu
+
+__all__ = ["AlbumSidebarContextMenu", "show_context_menu"]
